+++ src/DatePicker/Calendar/index.js
@@ -15,8 +15,6 @@ import './index.less'
 
 class Calendar extends Component {
 
-  static dayNames = ['一', '二', '三', '四', '五', '六', '日']
-
   constructor(props) {
     super()
     this.state = this.getDateState(props.date)
@@ -99,7 +97,7 @@ class Calendar extends Component {
     const isStart = timestrap === start
     const isEnd = timestrap === end
     const prefix = 'bfd-calendar__day--'
-    
+
     return classnames('bfd-calendar__day', {
       [`${prefix}today`]: timestrap === new Date().setHours(0, 0, 0, 0),
       [`${prefix}exclude`]: date.notThisMonth,
@@ -156,12 +154,12 @@ class Calendar extends Component {
     }
     return dates
   }
-  
+
   render() {
 
     const { currentYear, currentMonth } = this.state
-    const { start, end } = this.props
-    
+    const { start, end, captionRender, weekDayNames } = this.props
+
     const dates = this.getDates()
     const getComparerResult = this.disabledComparer()
 
@@ -173,13 +171,13 @@ class Calendar extends Component {
       <div className="bfd-calendar">
         <div className="bfd-calendar__header">
           <div className="bfd-calendar__header-left">
-            <Button 
+            <Button
               size="sm"
               icon="angle-double-left"
               transparent
               onClick={this.handleToggle.bind(this, -1, 'year')}
             />
-            <Button 
+            <Button
               size="sm"
               icon="angle-left"
               transparent
@@ -187,16 +185,16 @@ class Calendar extends Component {
             />
           </div>
           <span className="bfd-calendar__result">
-            {currentYear}年 {currentMonth + 1}月
+            {captionRender(this.state)}
           </span>
           <div className="bfd-calendar__header-right">
-            <Button 
+            <Button
               size="sm"
               icon="angle-right"
               transparent
               onClick={this.handleToggle.bind(this, 1)}
             />
-            <Button 
+            <Button
               size="sm"
               icon="angle-double-right"
               transparent
@@ -206,20 +204,20 @@ class Calendar extends Component {
         </div>
         <table>
           <thead>
-            <tr>{Calendar.dayNames.map((name, i) => <th key={i}>{name}</th>)}</tr>
+            <tr>{weekDayNames.map((name, i) => <th key={i}>{name}</th>)}</tr>
           </thead>
           <tbody>
             {Array(dates.length / 7 + 1).join(0).split('').map((v, i) => {
               return (
-                <tr key={i}>{Calendar.dayNames.map((name, j) => {
+                <tr key={i}>{weekDayNames.map((name, j) => {
                   const index = i * 7 + j
                   const date = dates[index]
                   return (
                     <td key={index}>
-                      <button 
+                      <button
                         type="button"
-                        disabled={getComparerResult(date)} 
-                        className={this.getDateClassNames(date, _start, _end)} 
+                        disabled={getComparerResult(date)}
+                        className={this.getDateClassNames(date, _start, _end)}
                         onClick={this.handleDaySelect.bind(this, date)}
                       >
                         {date.day}
@@ -227,12 +225,19 @@ class Calendar extends Component {
                     </td>
                   )
                 })}</tr>
-              ) 
+              )
             })}
           </tbody>
         </table>
       </div>
     )
+  }
+}
+
+Calendar.defaultProps = {
+  weekDayNames: ['一', '二', '三', '四', '五', '六', '日'],
+  captionRender(state) {
+    return `${state.currentYear}年 ${state.currentMonth + 1}月`
   }
 }
 
@@ -244,7 +249,9 @@ Calendar.propTypes = {
   max: checkDateTime,
   start: checkDateTime,
   end: checkDateTime,
-  onSelect: PropTypes.func
+  onSelect: PropTypes.func,
+  captionRender: PropTypes.func,
+  weekDayNames: PropTypes.array
 }
 
 export default Calendar