+++ CHANGELOG.md
@@ -1,15 +1,35 @@
 ## v1.4.0
 
-*2016-11-22*
+*2016-12-02*
 
-> `v1.4.0` 起，DropdownMenu 渲染到 body 节点内，相关组件 Select、MultipleSelect 展开层 DOM 位置均改变，自定义样式可能因 DOM 结构的改变而失效，请注意修改。
+> `v1.4.0` 起，DropdownMenu 渲染到 body 节点内，相关组件 Select, MultipleSelect 展开层 DOM 位置均改变，自定义的样式可能因 DOM 结构的改变而失效
 
 - Select onChange 新增参数 item
-- DropdownMenu 渲染至 body 节点内，相关组件 Select、MultipleSelect 展开层 DOM 位置均改变
+- DropdownMenu 渲染至 body 节点内，相关组件 Select, MultipleSelect 展开层 DOM 位置均改变
 - Dropdown 新增属性 aligned，实现 DropdownToggle 与 DropdownMenu 宽度相同
-- DropdownMenu 新增属性 direction、align，支持方向控制及对齐方式
-- Tooltip、TextOverflow 新增属性 align，支持对齐方式
+- DropdownMenu 新增属性 direction, align，支持方向控制及对齐方式
+- Tooltip, TextOverflow 新增属性 align，支持对齐方式
 - 修复 Percentage css 宽度无效的问题
+
+另外，新增大量属性支持自定义组件内部部分内容的显示，可用于国际化等需求，新增相关组件及属性如下
+
+- `Select`: searchPlaceholder, noOptionsContent, noMatchingContent
+- `CheckboxGroup`: toggleAllContent
+- `Paging`: auxiliaryRender
+- `MultipleSelect`: noOptionsContent, toggleAllContent
+- `DataTable`: noDataContent
+- `DataTable2`: noDataContent
+- `DatePicker/DateRange`: placeholder, captionRender, weekDayNames
+
+
+全局配置方式，以 Select placeholder 为例：
+```js
+import { Select } from 'bfd/Select'
+
+Object.assign(Select.defaultProps, {
+  placeholder: 'Please select'
+})
+```
 
 
 ## v1.3.4